Refactor other scraper to people_scraper and add logger.debug statements
