+++ app/people_scraper/extract.py
@@ -1,10 +1,11 @@
 from datetime import datetime
 
 import requests
+from flask import current_app
 
 from app.models import People, ProfessionalDetails, SideJobs
-from app.scraper.config import DETAILS_ENDPOINT, HEADERS, SEARCH_ENDPOINT
-from app.scraper.utils import (
+from app.people_scraper.config import DETAILS_ENDPOINT, HEADERS, SEARCH_ENDPOINT
+from app.people_scraper.utils import (
     find_request_verification_token,
     format_payload,
     professional_detail_already_exists,
@@ -20,17 +21,28 @@ def import_people_handler():
         HEADERS["__RequestVerificationToken"] = find_request_verification_token(
             r.content
         )
+        current_app.logger.debug(
+            f'Found CSRF token: {HEADERS["__RequestVerificationToken"]}'
+        )
+
         for search_string in search_strings():
             payload = format_payload(search_string)
             r = s.post(SEARCH_ENDPOINT, json=payload, headers=HEADERS)
-            print(r.status_code, r.url, payload)
+
+            current_app.logger.debug(
+                f"Collecting people from {r.url} with payload {payload}"
+            )
 
             if not r.ok:
-                print("error", r.status_code, r.content)
+                current_app.logger.error(
+                    f"Error during people collection: STATUS_CODE {r.status_code} | URL {r.url} | CONTENT {r.content}"
+                )
                 continue
 
             people = r.json().get("result", {}).get("model", {}).get("groupedItems", {})
-            print(f"Found {len(people)} people.")
+
+            current_app.logger.debug(f"{len(people)} people found for {r.url}")
+
             for person in people:
                 p_kwargs = People.from_dict(person)
                 person_already_exists = People.query.filter(
@@ -38,6 +50,10 @@ def import_people_handler():
                 ).all()
                 if not person_already_exists:
                     People.create(**p_kwargs)
+                else:
+                    current_app.logger.debug(
+                        f'Person with rechtspraak_id {p_kwargs.get("rechtspraak_id")} already exists'
+                    )
 
 
 def enrich_people_handler():
@@ -45,10 +61,14 @@ def enrich_people_handler():
 
     for person in people:
         r = requests.get(DETAILS_ENDPOINT + person.rechtspraak_id)
-        print(r.status_code, r.url)
+        current_app.logger.debug(
+            f"Enriching person {person.id} with information from {r.url}"
+        )
 
         if not r.ok:
-            print("error", r.status_code, r.content)
+            current_app.logger.error(
+                f"Error during person enrichment: {person.id}, {r.status_code}, {r.url}, {r.content}"
+            )
             continue
 
         person_json = r.json().get("model", {})
